Type t_dac_counter should have the same length as an element of pi_sp()
The pi_sp is an array of t_pi_sp_word, pi_sp'length will return the
size of the array, not the size in bits of t_pi_sp_word (what we
want). Change the t_dac_counter subtype declaration to use
pi_sp(0)'length instead. VHDL 2008 has a clearer syntax using the
'element attribute, but it might not be supported by some synthesis
tools.