Change PI output bias to 32767 to workaround implementation bug
The 32768 offset was chosen as a 'half-scale' value for the DACs, but
from my simulations it seems problematic (under certain conditions the
voltage output became frozen even when the current error was large).

The 32768 vs 32767 seems to suggest some signed arithmetic bug, but I
didn't investigated it any further.

It may fix an issue we observed in the field that sometimes after
loading a new gateware some fast correctors would saturate and the
only way to fix it was loading the gateware again.