Implement internal trigger generation from amplifier flags
This is useful to investigate the problems we are experiencing with
the power amplifiers entering in over-current protection mode without
actually experiencing an over-current condition. The internal trigger
can be routed to the acq core to obtain the current measurement and
voltage control at the instant of the event.

Be aware that the trigger will arrive with a variable latency that can
be after ~480 samples of the RTM-LAMP ADC due to the time necessary to
read the shift registers. For this motive, the acquisition should be
configured with an adequate number of pre-samples to actually observe
the current and voltage waveforms at the overcurrent / overtemperature
event.